refactor: organize builtin functions into modules
